requirements.txt
Flask==3.1.0
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gunicorn==23.0.0
python-dotenv==1.0.1
